+++ src/client/assets/images/chatImages/aristocrat_02_1.png
