+++ src/client/assets/images/chatImages/char_walk_right.gif
