+++ src/client/assets/images/chatImages/whiteCircle.png
